Add UserSessionLauncher to launch apps in user session
Introduces the UserSessionLauncher service to enable launching processes in the interactive user's session from a Windows service (Session 0). AppCatalogService is updated to use this new service for launching applications, ensuring UI apps can be started in the correct user context. Program.cs is updated to register UserSessionLauncher as a singleton.